resolve code bugs and code improvement
