+++ utils/utils.go
@@ -2,12 +2,10 @@ package utils
 
 import (
 	"encoding/json"
-	"fmt"
 	"log"
 	"net/http"
 	"net/mail"
 	"os"
-	"reflect"
 
 	"go.mongodb.org/mongo-driver/bson"
 )
@@ -68,38 +66,12 @@ func MapToBson(data map[string]interface{}) bson.M {
 	return bson.M(data) // they have the same underlying type so type conversion is enough
 }
 
-// StructToMap converts a struct of any type to a map[string]inteface{} based on struct tags
-// The struct tag is used to decide which field is added to the map.
-// This function is useful when you want to convert a model struct to a map[string]interface{}
-// the tag name here should be `bson` and the value should be the name of the struct field
-func StructToMap(inStruct interface{}, tag string) (map[string]interface{}, error) {
+// StructToMap converts a struct of any type to a map[string]inteface{}
+func StructToMap(inStruct interface{}) (map[string]interface{}, error) {
 	out := make(map[string]interface{})
-
-	v := reflect.ValueOf(inStruct)
-
-	// if it is a pointer to a struct, extract the element
-	if v.Kind() == reflect.Ptr {
-		v = v.Elem()
-	}
-
-	// only accepts struct kind, any other kind is an error
-	if v.Kind() != reflect.Struct {
-		return nil, fmt.Errorf("StructToMap only accepts structs or pointer to structs: got %T", v)
-	}
-
-	typ := v.Type()
-
-	for i := 0; i < v.NumField(); i++ {
-		field := typ.Field(i)
-		if tagVal := field.Tag.Get(tag); !shouldOmitTag(tagVal) {
-			out[tagVal] = v.Field(i).Interface()
-		}
-	}
+	inrec, _ := json.Marshal(inStruct)
+	json.Unmarshal(inrec, &out)
 	return out, nil
-}
-
-func shouldOmitTag(tagVal string) bool {
-	return tagVal == "" || tagVal == "-"
 }
 
 func ParseJsonFromRequest(r *http.Request, v interface{}) error {
